requirements.txt
streamlit
pandas
scikit-learn
requests
googletrans==4.0.0-rc1
google-cloud-texttospeech
 
